requirements-full.txt
Flask==3.0.0
Werkzeug==3.0.1
tensorflow-cpu==2.16.1
opencv-python-headless==4.9.0.80
numpy==1.26.3
python-dotenv==1.0.0
gunicorn==21.2.0
